Refactor pipeline.
Refactored pipeline into a class to avoid passing many arguments in the function calls and make the code more clean. Changed `control_activations_path` to `model_layer_activations_path` in `src/model/concept_history.py`. The first path is the entry directory with control activations for all models and layers and the latter one is a directory for specific model - layer combination.