Pipeline improvements.
History is sorted before passing to LLM. This is not the most efficient approach, but since we quickly want to test how will this perform I decided to do so.
Directories with images now contain name of the concept.
Refactored configs dataclasses to a separate `config.py` file.
Changed `concept_history` to a public field and removed setter, moved formatting logic to `history_managing.py`.