Parametrize the layer.
Added `layer` as argument to `main.py` and modified `get_activations` logic of `ExplainedModel` to extract activations from selected layer.
`ExplainedModel` is initialized with `layer` parameter as the layer won't change during its lifetime.
Changed `load` method of the `Model` to return the model resources instead of assigning it to a field to allow for testing. This will need some future fixes for the `LanguageModel`.